requirements.txt
azure-functions==1.24.0
cloudpickle==3.1.2
et_xmlfile==2.0.0
joblib==1.5.3
MarkupSafe==3.0.3
numpy==1.23.5
openpyxl==3.1.5
pandas==1.5.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.5
